Refactor/update utility classes usage for better clarity
