--- collector/test-data-uploader/src/main/kotlin/com/buildkite/test/collector/android/util/Helpers.kt
@@ -1,9 +1,0 @@
-package com.buildkite.test.collector.android.util
-
-import java.util.UUID
-
-internal object Helpers {
-    fun generateUUIDString(): String {
-        return UUID.randomUUID().toString()
-    }
-}
